smol-crewai/requirements.txt
crewai
langchain-groq
python-dotenv 